+++ apps/hub/src/app/ai/WorkflowSummaryList/index.tsx
@@ -1,52 +1,11 @@
-import { useSearchParams } from "next/navigation";
-import { FC, use } from "react";
+import { FC } from "react";
 
 import { ClientWorkflow } from "@quri/squiggle-ai";
-import { Button, FireIcon } from "@quri/ui";
 
-import { AdminContext } from "@/components/admin/AdminProvider";
 import { LoadMoreViaSearchParam } from "@/components/LoadMoreViaSearchParam";
-import { useUpdateSearchParams } from "@/lib/hooks/useUpdateSearchParams";
 
+import { WorkflowListAdminControls } from "./WorkflowListAdminControls";
 import { WorkflowSummaryItem } from "./WorkflowSummaryItem";
-
-const WorkflowListAdminControls = () => {
-  const { isAdminMode } = use(AdminContext);
-  const updateSearchParams = useUpdateSearchParams();
-  const searchParams = useSearchParams();
-
-  if (!isAdminMode) {
-    return null;
-  }
-
-  return searchParams.get("allUsers") ? (
-    <Button
-      size="small"
-      onClick={() =>
-        updateSearchParams((params) => {
-          params.delete("allUsers");
-          return params;
-        })
-      }
-    >
-      <FireIcon className="h-4 w-4 text-red-500" />
-      Show my
-    </Button>
-  ) : (
-    <Button
-      size="small"
-      onClick={() =>
-        updateSearchParams((params) => {
-          params.set("allUsers", "true");
-          return params;
-        })
-      }
-    >
-      <FireIcon className="h-4 w-4 text-red-500" />
-      Show all
-    </Button>
-  );
-};
 
 export const WorkflowSummaryList: FC<{
   workflows: ClientWorkflow[];
